DEV/TEST: Implement further elements of the parser
New Enums for the numeric token fields, and the number/string
match styles (capture or not for strings; various for numbers)

Revise Enum values to the single-character flags for those
types of values.

Begin implementation of pyparsing as parser for the mini-language
itself, even though the underlying search is run via regex.
Probably for the best, as the regex engine appears to be
*SUBSTANTIALLY* faster than pyparsing. Makes sense, if pyparsing
is pure Python.

Parser test now is actually a real test of the syntax! (Or, well,
PART of the syntax, anyways...)