DEV/TEST: Start attempting no-space gap; add tests
Various tests for captures, non-captures, etc. added.

Started trying to implement the no-space-gap feature. The original
idea of just not inserting the preceding space pattern bit is
not going to work, because it runs afoul of both the preceding
and the 'x'-marked tokens being wordified.

This may require a complete first pass through all the tokens
to identify any that are no-spaced, followed by a second pass
to actually assemble the patterns, with wordification
applied appropriately to no-space groups.