REFACTOR: Implement Token parsing class
The mini-language token parsing is now encapsulated in Token,
in parser.py.  This should make for a much cleaner scan
over the mini-language content within Parser, and should allow
for much more flexible adaptation/augmentation/modification
of the regex patterns generated out of each token/Token.

Add PentError and BadTokenError custom exceptions.

Add Enum for 'any' match types. Will be useful if/when a
'capturing-any' variant is implemented, but was mainly added
to allow use of the 'parser._has_value' helper function.