DEV: Complete rework of token system
But now it's a LOT more general, and should hopefully avoid
all of the ugliness of the <..< >..> preceding/following literals
hanging on the number types.

The tokens are still bulkier than I'd like, but they're not TOO
bad. Probably quite readable once one gets used to the syntax?

Closes #14. Also closes #9, with the expansion of the 'any' type
to both capturing and non-capturing variants.

Also closes #6 by obviating it, since this is now best handled
by no-space flags and string literals.