TDD: Add 1-D parser test; partial implement
Add an example ORCA .hess file, for use in testing the higher-level
Parser functionality.

Modify Parser.convert_line to also return the value of the group_id,
as it will be necessary to increment and re-pass this value when
building with-capture patterns for head/tail containing multiple
lines. (This necessitates [0]-indexing the return value from
.convert_line if all you care about is the regex pattern.)

Implement a first pass on the total-pattern construction @property
of Parser. Added test seems to confirm that the 1-D construction
works properly. Remains to be seen if the 2-D construction also
works (#34).  Should possibly try to develop at least a toy 3-D dataset
and see how it parses (#35).

For both 1-D and 2-D (and generalized 3-D .. n-D), need to implement
the structured data retrieval and cross-check against sample data (#36).

Lots of sample data.