DEV/TEST: Add ThruList in capture_struct
In many situations, only a single captured block can be expected,
in which case having to index to that first element will lead to
a lot of "[0]" cluttering up the code.

ThruList avoids that by automatically indexing into the object
at item 0 if the passed key isn't an int.

User can always explicitly index the ThruList[0] if there's a
possibility for more than one result.

Decent suite added.

[skip ci]