TEST: Complete CAS deep-structure tests
At least, for the CI block info test. Could probably add for
the CI state info test, too, but it gets the point across.

Now to add some for other quantities ... geom opts in ORCA,
maybe SCF in ORCA and GAMESS .. definitely lots of stuff from
Multiwfn. Progress has been made on #41, though.

At this point, the two major items of #48 are implemented; leaving
it open for now in case more data formats come to mind or are
suggested. Accordingly, this commit closes #37.

Believe #47 is successfully implemented at this point, but tests
are needed to be sure that head/tail do actually handle Parsers
correctly.

Closes #44 since `capture_head` and `capture_tail` have been
removed.