REFAC: Raw-ify all \n strings; update skip msgs
Probably better practice to have escaped newlines in the
generated regex, rather than explicit newline literals. So,
constant strings with newlines were changed to raw-strings.

The skip messages on the tests for optional/zero-or-more tokens
were somewhat out of date. They now reflect the plan not to
ever attempt to implement them, since the regex pattern has to be
different depending on whether the optional/z.o.m. value is actually
present in the text or not.

[skip ci]