DEV: Make entire sections possibly optional
If all of the string patterns in a given head or tail are optional,
add a '?' to that capture group in case there's no content there
for a given (sub)match. Otherwise, re.finditer et al. won't match
correctly when the block is missing, even if they should.

Now, need to adjust the code in capture_str_pattern to account
for the possibility that 'text' could be None, in those situations
where the entire section is absent for a given (sub)match. In this
case, since the capture group itself is optional, the regex parser
adds the group to its index but fills the .group with None if
the match isn't found.

[skip ci]