TEST: Revise opt status test w/o optional
Works fine when the similar formatting of the data lines is exploited
and it's just an arbitrary number of repeated lines in the body.

Not perfect/ideal, but workable.

While not as robust an implementation as hoped, closes #43.
Task remains (per #51 and #52) to explore and document the
limitations of the current implementation.