TDD: Add column_stack_2d, various live datasets
Helper function column_stack_2d was pretty simple to put together.
It will probably fail in a lot of different circumstances, and
probably in ways that might appear to be success but instead give
unexpected results. #58 created to plant the flag that more
checking is probably needed.

Also add a bunch of live-data test files.