TEST: Add mwfn LI/DI livedata tests
THEY WORK.

Data can be pulled in as list/matrix (of values as str) with
three steps:

 1. Retrieve the text to be parsed
 2. Define Parser
 3. Execute the capture

A "4th step" to post-convert the structure of the data, or to
screen out extraneous content, may be required, but it shouldn't
be too hard to implement in most cases.

Checks a box in #41.  Should also supply coverage for the
one-or-more case of the 'misc content' token (&+).