TEST/ADMIN: Draft README, add test; add numpy
For the documentation testing, going to want numpy available since
most users are expected to want the data in numpy arrays at the end.
So, numpy added to requirements-dev. Also added to -travis; should
be fine not to have it in -rtd, since not running doctest on that build.

Draft an initial README, so that there's *some*thing there for 0.2.
Add doctest to suite to test the examples there.