Reorganize headings and tokens
Suggest removing the introduction of the actual tokens from the bold headings. When I run my eyes over the page, it is easy to miss the token. The first time I scanned through, I found myself wonder "wait.. where is the token? what is it...?" and then scanning in reverse until realizing, "oh, it was in the heading."